move status filter from profile configuration to task
