+++ src/main/java/com/aspectsecurity/contrast/contrastjenkins/VulnerabilityTrendRecorder.java
@@ -91,8 +91,8 @@ public class VulnerabilityTrendRecorder extends Recorder {
                     filterForm.setVulnTypes(Collections.singletonList(condition.getThresholdVulnType()));
                 }
 
-                if (!profile.getVulnerabilityStatuses().isEmpty()) {
-                    filterForm.setStatus(profile.getVulnerabilityStatuses());
+                if (!condition.getVulnerabilityStatuses().isEmpty()) {
+                    filterForm.setStatus(condition.getVulnerabilityStatuses());
                 }
 
                 traces = contrastSDK.getTracesInOrg(profile.getOrgUuid(), filterForm);
