+++ src/main/java/com/aspectsecurity/contrast/contrastjenkins/VulnerabilityTrendStep.java
@@ -203,9 +203,6 @@ public class VulnerabilityTrendStep extends AbstractStepImpl {
                 if (step.getRule() != null) {
                     filterForm.setVulnTypes(Collections.singletonList(step.getRule()));
                 }
-                if (!teamServerProfile.getVulnerabilityStatuses().isEmpty()) {
-                    filterForm.setStatus(teamServerProfile.getVulnerabilityStatuses());
-                }
 
                 traces = contrastSDK.getTracesInOrg(teamServerProfile.getOrgUuid(), filterForm);
             } catch (Exception e) {
