+++ src/main/resources/com/aspectsecurity/contrast/contrastjenkins/VulnerabilityTrendRecorder/config.jelly
@@ -9,20 +9,54 @@
         <f:entry description="Conditions for verifying your build">
             <f:repeatable field="conditions">
                 <table width="100%">
-                    <f:entry title="Count" field="thresholdCount" help="/plugin/contrast-continuous-application-security/help-thresholdCount.html">
-                        <f:number name="thresholdCount" field="thresholdCount" default="0" clazz="required positive-number" />
+                    <f:entry title="Count" field="thresholdCount"
+                             help="/plugin/contrast-continuous-application-security/help-thresholdCount.html">
+                        <f:number name="thresholdCount" field="thresholdCount" default="0"
+                                  clazz="required positive-number"/>
                     </f:entry>
 
-                    <f:entry title="Severity" field="thresholdSeverity" help="/plugin/contrast-continuous-application-security/help-thresholdSeverity.html">
+                    <f:entry title="Severity" field="thresholdSeverity"
+                             help="/plugin/contrast-continuous-application-security/help-thresholdSeverity.html">
                         <f:select/>
                     </f:entry>
 
-                    <f:entry title="Vulnerability Type" field="thresholdVulnType" help="/plugin/contrast-continuous-application-security/help-thresholdVulnType.html">
+                    <f:entry title="Vulnerability Type" field="thresholdVulnType"
+                             help="/plugin/contrast-continuous-application-security/help-thresholdVulnType.html">
                         <f:select/>
                     </f:entry>
 
-                    <f:entry title="Application Name" help="/plugin/contrast-continuous-application-security/help-applicationName.html">
+                    <f:entry title="Application Name"
+                             help="/plugin/contrast-continuous-application-security/help-applicationName.html">
                         <f:textbox name="applicationName" field="applicationName" default="${it.displayName}"/>
+                    </f:entry>
+
+                    <f:entry title="Vulnerability statuses" help="/plugin/contrast-continuous-application-security/help-vulnerabilityStatus.html">
+                        <f:checkbox title="Auto-Remediated" name="autoRemediated" field="autoRemediated"/>
+                    </f:entry>
+
+                    <f:entry>
+                        <f:checkbox title="Not a Problem" name="notAProblem" field="notAProblem"/>
+                    </f:entry>
+                    <f:entry>
+                        <f:checkbox title="Fixed" name="fixed" field="fixed"/>
+                    </f:entry>
+                    <f:entry>
+                        <f:checkbox title="Confirmed" name="confirmed" field="confirmed"/>
+                    </f:entry>
+                    <f:entry>
+                        <f:checkbox title="Remediated" name="remediated" field="remediated"/>
+                    </f:entry>
+                    <f:entry>
+                        <f:checkbox title="Being Tracked" name="beingTracked" field="beingTracked"/>
+                    </f:entry>
+                    <f:entry>
+                        <f:checkbox title="Suspicious" name="suspicious" field="suspicious"/>
+                    </f:entry>
+                    <f:entry>
+                        <f:checkbox title="Reported" name="reported" field="reported"/>
+                    </f:entry>
+                    <f:entry>
+                        <f:checkbox title="Untracked" name="untracked" field="untracked"/>
                     </f:entry>
 
                     <f:entry title="">
