+++ src/test/java/com/aspectsecurity/contrast/contrastjenkins/ThresholdConditionStub.java
@@ -4,7 +4,9 @@ package com.aspectsecurity.contrast.contrastjenkins;
 public class ThresholdConditionStub extends ThresholdCondition {
 
     public ThresholdConditionStub() {
-        super(0, "test", "test", "test");
+        super(0, "test", "test", "test",
+                false, false,false, false, false, false,
+                false, false, false);
     }
 
     public static class ThresholdConditionDescriptorStub extends ThresholdCondition.DescriptorImpl {
