+++ src/test/java/com/aspectsecurity/contrast/contrastjenkins/VulnerabilityTrendRecorderTest.java
@@ -75,9 +75,7 @@ public class VulnerabilityTrendRecorderTest {
         given(VulnerabilityTrendHelper.createSDK(anyString(), anyString(), anyString(), anyString())).willReturn(contrastSDKMock);
 
         TeamServerProfile profile = new TeamServerProfile("local", "contrast", "demo", "demo",
-                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString(),
-                false, false, false, false, false,
-                false, false, false, false);
+                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString());
 
         given(VulnerabilityTrendHelper.getProfile(anyString())).willReturn(profile);
 
@@ -125,9 +123,7 @@ public class VulnerabilityTrendRecorderTest {
         given(VulnerabilityTrendHelper.createSDK(anyString(), anyString(), anyString(), anyString())).willReturn(contrastSDKMock);
 
         TeamServerProfile profile = new TeamServerProfile("local", "contrast", "demo", "demo",
-                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString(),
-                false, false, false, false, false,
-                false, false, false, false);
+                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString());
 
         given(VulnerabilityTrendHelper.getProfile(anyString())).willReturn(profile);
 
@@ -175,9 +171,7 @@ public class VulnerabilityTrendRecorderTest {
         given(VulnerabilityTrendHelper.createSDK(anyString(), anyString(), anyString(), anyString())).willReturn(contrastSDKMock);
 
         TeamServerProfile profile = new TeamServerProfile("local", "contrast", "demo", "demo",
-                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString(),
-                false, false, false, false, false,
-                false, false, false, false);
+                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString());
 
         given(VulnerabilityTrendHelper.getProfile(anyString())).willReturn(profile);
 
@@ -227,9 +221,7 @@ public class VulnerabilityTrendRecorderTest {
         given(VulnerabilityTrendHelper.createSDK(anyString(), anyString(), anyString(), anyString())).willReturn(contrastSDKMock);
 
         TeamServerProfile profile = new TeamServerProfile("local", "contrast", "demo", "demo",
-                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString(),
-                false, false, false, false, false,
-                false, false, false, false);
+                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString());
 
         given(VulnerabilityTrendHelper.getProfile(anyString())).willReturn(profile);
 
@@ -269,6 +261,13 @@ public class VulnerabilityTrendRecorderTest {
         when(thresholdCondition.getThresholdCount()).thenReturn(0);
         when(thresholdCondition.getThresholdSeverity()).thenReturn("test");
         when(thresholdCondition.getThresholdVulnType()).thenReturn("test");
+
+        List<String> status = new ArrayList();
+        status.add(Constants.VULNERABILITY_STATUS_AUTO_REMEDIATED);
+        status.add(Constants.VULNERABILITY_STATUS_CONFIRMED);
+        status.add(Constants.VULNERABILITY_STATUS_SUSPICIOUS);
+        when(thresholdCondition.getVulnerabilityStatuses()).thenReturn(status);
+
         conditions.add(thresholdCondition);
 
         VulnerabilityTrendRecorder vulnerabilityTrendRecorder = spy(new VulnerabilityTrendRecorder(conditions, "test"));
@@ -283,9 +282,7 @@ public class VulnerabilityTrendRecorderTest {
         given(VulnerabilityTrendHelper.createSDK(anyString(), anyString(), anyString(), anyString())).willReturn(contrastSDKMock);
 
         TeamServerProfile profile = new TeamServerProfile("local", "contrast", "demo", "demo",
-                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString(),
-                true, true, true, false, false,
-                false, false, false, false);
+                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString());
 
         given(VulnerabilityTrendHelper.getProfile(anyString())).willReturn(profile);
