+++ src/test/java/com/aspectsecurity/contrast/contrastjenkins/VulnerabilityTrendStepTest.java
@@ -74,9 +74,7 @@ public class VulnerabilityTrendStepTest {
         given(VulnerabilityTrendHelper.createSDK(anyString(), anyString(), anyString(), anyString())).willReturn(contrastSDKMock);
 
         TeamServerProfile profile = new TeamServerProfile("local", "contrast", "demo", "demo",
-                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString(),
-                false, false, false, false, false,
-                false, false, false, false);
+                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString());
 
         given(VulnerabilityTrendHelper.getProfile(anyString())).willReturn(profile);
 
@@ -107,9 +105,7 @@ public class VulnerabilityTrendStepTest {
         given(VulnerabilityTrendHelper.createSDK(anyString(), anyString(), anyString(), anyString())).willReturn(contrastSDKMock);
 
         TeamServerProfile profile = new TeamServerProfile("local", "contrast", "demo", "demo",
-                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString(),
-                false, false, false, false, false,
-                false, false, false, false);
+                "www.google.com", "org-uuid", "Jenkins", false, Result.FAILURE.toString());
 
         given(VulnerabilityTrendHelper.getProfile(anyString())).willReturn(profile);
